requirements.txt
-e git+https://github.com/mgax/chaussette.git@301e4baa2b5251e780a66b37c4478d4b51776c80#egg=chaussette-dev
-e git+https://github.com/mgax/flask-browserid.git@8e610c302bca5336076c13278005456fc40df3fd#egg=Flask_BrowserID-dev
-e git+https://github.com/mgax/raven-python.git@0b3a3e74ee6a21291d2fa59bcc6ee161d25c0b27#egg=raven-dev
-e git+https://github.com/mitsuhiko/flask-sqlalchemy.git@2edb3f14931eefde4bf467e88f7373704ce1ecd3#egg=Flask_SQLAlchemy-dev
-e git+https://github.com/mitsuhiko/flask.git@2e17ad7fd42d8575f308af9c17496d38ceb72b5e#egg=Flask-dev
-e git+https://github.com/mrjoes/flask-admin.git@ecaa5d2c49aa63a12f40fa50186aa36b3b8712ab#egg=Flask_Admin-dev
-e git+https://github.com/techniq/flask-script.git@6e47afcea25e2ad26a9c32371b3ebe34f3704b12#egg=Flask_Script-dev
blinker==1.3
cssselect==0.8
Flask-Login==0.2.6
Flask-WTF==0.8.4
itsdangerous==0.22
Jinja2==2.7
lxml==3.2.1
MarkupSafe==0.18
path.py==4.3
pyquery==1.2.4
requests-cache==0.4.0
requests==1.2.3
six==1.3.0
SQLAlchemy==0.8.2
waitress==0.8.5
Werkzeug==0.9.2
WTForms==1.0.4
